Rename command-x modules to commands-x
Seems slightly clearer
