+++ src/test/browser/test-command.js
@@ -4,7 +4,7 @@ const {cmpNode} = require("../cmp")
 const {doc, blockquote, pre, h1, h2, p, li, ol, ul, em, hr} = require("../build")
 
 const commands = require("../../commands")
-const listCommands = require("../../command-list")
+const listCommands = require("../../commands-list")
 const {Schema} = require("../../model")
 const {schema, Heading, Doc} = require("../../schema-basic")
